Move TomographyExperiment & co to experiment/_main.py
